feat: meaningful verify meessages
